+++ app/src/main/java/com/example/pilotlogbook/data/room/entities/dailyflight/tuples/OperationalConditionTimeTuple.kt
@@ -3,6 +3,6 @@ package com.example.pilotlogbook.data.room.entities.dailyflight.tuples
 import androidx.room.ColumnInfo
 
 data class OperationalConditionTimeTuple(
-    @ColumnInfo(name = "operation_night")val night: Double?,
-    val ifr: Double?
+    @ColumnInfo(name = "operation_night")val night: Long?,
+    val ifr: Long?
 )
