+++ app/src/main/java/com/example/pilotlogbook/data/room/entities/dailyflight/tuples/PilotFunctionTimeTuple.kt
@@ -1,8 +1,8 @@
 package com.example.pilotlogbook.data.room.entities.dailyflight.tuples
 
 data class PilotFunctionTimeTuple(
-    val pilotInComand: Double?,
-    val coPilot: Double?,
-    val dual: Double?,
-    val instructor: Double?
+    val pilotInComand: Long?,
+    val coPilot: Long?,
+    val dual: Long?,
+    val instructor: Long?
 )
