+++ app/src/main/java/com/example/pilotlogbook/data/room/entities/dailyflight/tuples/SinglePilotTimeTuple.kt
@@ -1,6 +1,6 @@
 package com.example.pilotlogbook.data.room.entities.dailyflight.tuples
 
 data class SinglePilotTimeTuple(
-    val se: Double?,
-    val me: Double?
+    val se: Long?,
+    val me: Long?
 )
