+++ app/src/main/java/com/example/pilotlogbook/data/room/entities/dailyflight/tuples/SyntheticTrainingDevicesSessionTuple.kt
@@ -5,5 +5,5 @@ import androidx.room.ColumnInfo
 data class SyntheticTrainingDevicesSessionTuple(
     @ColumnInfo(name = "synthetic_date")val date: String?,
     val type: String?,
-    val totalTimeOfSession: Double?
+    val totalTimeOfSession: Long?
 )
